fix: move evalhub kueue tests under model_explainability folder
