+++ tests/model_explainability/evalhub/evalhub_kueue_integration/scripts/cleanup_kueue_resources.sh
