+++ tests/model_explainability/evalhub/evalhub_kueue_integration/test_negative.py
@@ -4,7 +4,7 @@ import pytest
 from kubernetes.dynamic import DynamicClient
 from ocp_resources.namespace import Namespace
 
-from tests.eval_hub.evalhub_kueue_integration.constants import (
+from tests.model_explainability.evalhub.evalhub_kueue_integration.constants import (
     CLUSTER_QUEUE_NAME,
     DEFAULT_CPU_QUOTA,
     DEFAULT_MEMORY_QUOTA,
@@ -12,7 +12,7 @@ from tests.eval_hub.evalhub_kueue_integration.constants import (
     RESOURCE_FLAVOR_NAME,
     EvalJobState,
 )
-from tests.eval_hub.evalhub_kueue_integration.utils import (
+from tests.model_explainability.evalhub.evalhub_kueue_integration.utils import (
     delete_eval_job,
     get_eval_job,
     submit_eval_job,
