+++ tests/model_explainability/evalhub/evalhub_kueue_integration/test_preemption.py
@@ -5,14 +5,14 @@ from kubernetes.dynamic import DynamicClient
 from ocp_resources.job import Job
 from ocp_resources.namespace import Namespace
 
-from tests.eval_hub.evalhub_kueue_integration.constants import (
+from tests.model_explainability.evalhub.evalhub_kueue_integration.constants import (
     KUEUE_QUEUE_NAME_LABEL,
     LOCAL_QUEUE_NAME,
     RESOURCE_FLAVOR_NAME,
     SMALL_CPU_QUOTA,
     SMALL_MEMORY_QUOTA,
 )
-from tests.eval_hub.evalhub_kueue_integration.utils import (
+from tests.model_explainability.evalhub.evalhub_kueue_integration.utils import (
     delete_eval_job,
     submit_eval_job,
     wait_for_job_running_or_completed,
