+++ tests/model_explainability/evalhub/evalhub_kueue_integration/test_priority_scheduling.py
@@ -4,7 +4,7 @@ import pytest
 from kubernetes.dynamic import DynamicClient
 from ocp_resources.namespace import Namespace
 
-from tests.eval_hub.evalhub_kueue_integration.constants import (
+from tests.model_explainability.evalhub.evalhub_kueue_integration.constants import (
     CLUSTER_QUEUE_NAME,
     LOCAL_QUEUE_NAME,
     RESOURCE_FLAVOR_NAME,
@@ -12,7 +12,7 @@ from tests.eval_hub.evalhub_kueue_integration.constants import (
     SMALL_MEMORY_QUOTA,
     EvalJobState,
 )
-from tests.eval_hub.evalhub_kueue_integration.utils import (
+from tests.model_explainability.evalhub.evalhub_kueue_integration.utils import (
     delete_eval_job,
     get_eval_job,
     submit_eval_job,
