+++ tests/model_explainability/evalhub/evalhub_kueue_integration/test_upgrade.py
@@ -3,14 +3,14 @@
 import pytest
 from ocp_resources.namespace import Namespace
 
-from tests.eval_hub.evalhub_kueue_integration.constants import (
+from tests.model_explainability.evalhub.evalhub_kueue_integration.constants import (
     CLUSTER_QUEUE_NAME,
     DEFAULT_CPU_QUOTA,
     DEFAULT_MEMORY_QUOTA,
     LOCAL_QUEUE_NAME,
     RESOURCE_FLAVOR_NAME,
 )
-from tests.eval_hub.evalhub_kueue_integration.utils import (
+from tests.model_explainability.evalhub.evalhub_kueue_integration.utils import (
     delete_eval_job,
     get_health,
     submit_eval_job,
